new:64bit向けビルドに対応

+++ kernel/proc.h
@@ -10,40 +10,40 @@
 // コンテクストスイッチのために保存するレジスタ群.
 struct context
 {
-    uint32_t ra;   // x1: return address
-    uint32_t sp;   // x2: stack pointer
-    uint32_t gp;   // x3: global pointer
-    uint32_t tp;   // x4: thread pointer
-    uint32_t t0;   // x5
-    uint32_t t1;   // x6
-    uint32_t t2;   // x7
-    uint32_t s0;   // x8 (frame pointer)
-    uint32_t s1;   // x9
-    uint32_t a0;   // x10
-    uint32_t a1;   // x11
-    uint32_t a2;   // x12
-    uint32_t a3;   // x13
-    uint32_t a4;   // x14
-    uint32_t a5;   // x15
-    uint32_t a6;   // x16
-    uint32_t a7;   // x17
-    uint32_t s2;   // x18
-    uint32_t s3;   // x19
-    uint32_t s4;   // x20
-    uint32_t s5;   // x21
-    uint32_t s6;   // x22
-    uint32_t s7;   // x23
-    uint32_t s8;   // x24
-    uint32_t s9;   // x25
-    uint32_t s10;  // x26
-    uint32_t s11;  // x27
-    uint32_t t3;   // x28
-    uint32_t t4;   // x29
-    uint32_t t5;   // x30
-    uint32_t t6;   // x31
+    uint64_t ra;   // x1: return address
+    uint64_t sp;   // x2: stack pointer
+    uint64_t gp;   // x3: global pointer
+    uint64_t tp;   // x4: thread pointer
+    uint64_t t0;   // x5
+    uint64_t t1;   // x6
+    uint64_t t2;   // x7
+    uint64_t s0;   // x8 (frame pointer)
+    uint64_t s1;   // x9
+    uint64_t a0;   // x10
+    uint64_t a1;   // x11
+    uint64_t a2;   // x12
+    uint64_t a3;   // x13
+    uint64_t a4;   // x14
+    uint64_t a5;   // x15
+    uint64_t a6;   // x16
+    uint64_t a7;   // x17
+    uint64_t s2;   // x18
+    uint64_t s3;   // x19
+    uint64_t s4;   // x20
+    uint64_t s5;   // x21
+    uint64_t s6;   // x22
+    uint64_t s7;   // x23
+    uint64_t s8;   // x24
+    uint64_t s9;   // x25
+    uint64_t s10;  // x26
+    uint64_t s11;  // x27
+    uint64_t t3;   // x28
+    uint64_t t4;   // x29
+    uint64_t t5;   // x30
+    uint64_t t6;   // x31
 
-    uint32_t pc;   // program counter
-    uint32_t mstatus; // machine status register (if必要)
+    uint64_t pc;   // program counter
+    uint64_t mstatus; // machine status register (if必要)
 };
 
 // プロセス状態
@@ -61,10 +61,10 @@ enum procstate
 // 割り込みや例外で自動待避するCSRやレジスタ群を保存する
 struct trapframe
 {
-    uint32_t mepc;
-    uint32_t mstatus;
-    uint32_t mcause;
-    uint32_t mtval;
+    uint64_t mepc;
+    uint64_t mstatus;
+    uint64_t mcause;
+    uint64_t mtval;
 };
 
 // プロセス制御構造体
